requirements.txt
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.0.1
